requirements.txt
bentoml[grpc]==1.1.7
cattrs==23.1.2
torch==2.1.2
librosa==0.9.0
inflect
git+https://github.com/NVIDIA/dllogger@v1.0.0#egg=dllogger
tensorboardX==2.0
tensorboard==2.15.1
tritonclient[http]==2.39.0
fastapi==0.96.0
loguru==0.7.2
uvicorn[standard]==0.22.0
onnxruntime
feast[redis,postgres]==0.34.1
python-dotenv==1.0.0
psycopg2-binary==2.9.7
pyspark==3.0.1
kfp==1.8.3